mesafe değişkeninin veri tipini int den double a çevirdim
Java ile gidilen mesafeye(KM) göre taksimetre tutarını ekrana yazdıran bir program yazın.
Oluşabilecek hataların önüne geçmek için

+++ Taksimetre.java
@@ -10,7 +10,7 @@ public class Taksimetre {
         Scanner input = new Scanner(System.in);
         double kmCarpani = 2.20;
         System.out.println("Mesafe Giriniz(KM cinsinden) :");
-        int mesafe = input.nextInt();
+        double mesafe = input.nextDouble();
 
         double ucret = (10 + (mesafe * kmCarpani)) < 20 ? 20 : (10 + (mesafe * kmCarpani));
          ucret = Math.round(ucret);
